feat: add router and not found page
